Move license helpers to new file
