+++ src/utils/general-helpers.ts
@@ -32,29 +32,6 @@ export const getPeriodString = (from: string | undefined, to: string | undefined
 
 export const getIdentifierFromId = (id: string) => id.split('/').pop() ?? '';
 
-export const equalUris = (uri1: string | null, uri2: string | null) => {
-  if (!uri1 || !uri2) {
-    return false;
-  }
-  if (uri1 === uri2) {
-    return true;
-  }
-  const urlObj1 = new URL(uri1);
-  const urlObj2 = new URL(uri2);
-
-  if (urlObj1.hostname === urlObj2.hostname) {
-    if (urlObj1.pathname === urlObj2.pathname) {
-      return true;
-    }
-  }
-  return (
-    urlObj1.hostname === urlObj2.hostname &&
-    removeTrailingSlash(urlObj1.pathname) === removeTrailingSlash(urlObj2.pathname)
-  );
-};
-
-const removeTrailingSlash = (value: string) => (value.endsWith('/') ? value.slice(0, -1) : value);
-
 export const getInitials = (name: string) => {
   if (!name) return '';
